BE/requirements.txt
fastapi
uvicorn
python-jose
python-multipart
pydantic
passlib
sqlalchemy

Pillow
numpy

# CORS
starlette

# Transformers / CLIP
transformers
torch
torchvision
torchaudio

# For visualization
matplotlib
seaborn

# Optional but recommended
python-dotenv
